+++ lib/cloudflare_rails/remote_ip_proxies.rb
@@ -1,0 +1,9 @@
+module CloudflareRails
+  # patch ActionDispatch::RemoteIP to use our cloudflare ips - this way
+  # request.remote_ip is correct inside of rails
+  module RemoteIpProxies
+    def proxies
+      super + Importer.cloudflare_ips
+    end
+  end
+end