+++ spec/cloudflare/rails_spec.rb
@@ -10,7 +10,7 @@ describe CloudflareRails do
       let!(:rails_app) do
         ::ActiveSupport::Dependencies.autoload_once_paths = []
         ::ActiveSupport::Dependencies.autoload_paths = []
-        Class.new(::Rails::Application) do
+        Class.new(Rails::Application) do
           config.active_support.deprecation = :stderr
           config.eager_load = false
           config.cache_store = :null_store
@@ -81,14 +81,14 @@ describe CloudflareRails do
       if ENV['RACK_ATTACK']
         it "monkey-patches rack-attack" do
           rails_app.initialize!
-          expect(Rack::Attack::Request.included_modules).to include(CloudflareRails::Railtie::CheckTrustedProxies)
+          expect(Rack::Attack::Request.included_modules).to include(CloudflareRails::CheckTrustedProxies)
         end
       end
 
       it "works with valid responses" do
         expect_any_instance_of(Logger).not_to receive(:error)
         rails_app.initialize!
-        expect(Set.new(CloudflareRails::Railtie::Importer.cloudflare_ips(refresh: true))).
+        expect(Set.new(CloudflareRails::Importer.cloudflare_ips(refresh: true))).
           to eq(Set.new((ips_v4_body + ips_v6_body).split("\n").map { |ip| IPAddr.new ip }))
       end
 
@@ -99,7 +99,7 @@ describe CloudflareRails do
         it "doesn't break but still logs the error" do
           expect_any_instance_of(Logger).to receive(:error).once.and_call_original
           rails_app.initialize!
-          expect(CloudflareRails::Railtie::Importer.cloudflare_ips(refresh: true)).to be_blank
+          expect(CloudflareRails::Importer.cloudflare_ips(refresh: true)).to be_blank
         end
       end
 
@@ -110,7 +110,7 @@ describe CloudflareRails do
         it "doesn't break but still logs the error" do
           expect_any_instance_of(Logger).to receive(:error).once.and_call_original
           rails_app.initialize!
-          expect(CloudflareRails::Railtie::Importer.cloudflare_ips(refresh: true)).to be_blank
+          expect(CloudflareRails::Importer.cloudflare_ips(refresh: true)).to be_blank
         end
       end